test(frontend): reuse HTTP service suite for zero-top logprobs
